+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+ - Change `subset_shape` in `zarrsArrayGetSubsetSize` to `*u64`
+
 ## [0.5.1] - 2024-03-18
 
 ### Changed